Added files via upload
This is a basic class using a switch to define the policy of insurance and the options we agreed on before the break, it does currently not work in tandem with the DB or server but gives a clear layout. Due to my laptop not correctly having the set up i just want to show that work and planning is taking place. 

Thanks